Moved all form related data out into its own package. Edited setup and imports accordingly.
