+++ uchicagoldrRecords/digitalAcquisitionRead.py
