+++ uchicagoldrRecords/mapTemplate.txt
